Updates Start view
